Refactor visitor and unwrappable
